Add files via upload
The “Bank Account Management System” project is a model Internet Banking
Site. This site enables the customers to perform the basic banking transactions by
sitting at their office or at homes through PC or laptop. The system provides the
access to the customer to create an account, deposit/withdraw the cash from his
account, also to view reports of all accounts present. The customers can access the
banks website for viewing their Account details and perform the transactions on
account as per their requirements. With Internet Banking, the brick and mortar
structure of the traditional banking gets converted into a click and portal model,
thereby giving a concept of virtual banking a real shape. Thus, today's banking is
no longer confined to branches. E-banking facilitates banking transactions by
customers round the clock globally.
The primary aim of this “Bank Account Management System” is to provide an
improved design methodology, which envisages the future expansion, and
modification, which is necessary for a core sector like banking. This necessitates
the design to be expandable and modifiable and so a modular approach is used in
developing the application software.
Anybody who is an Account holder in this bank can become a member of Bank
Account
Management System. He has to fill a form with his personal details and Account
Number.
Bank is the place where customers feel the sense of safety for their property. In the
bank, customers deposit and withdraw their money. Transaction of money also is a
part where customer takes shelter of the bank. Now to keep the belief and trust of
customers, there is the positive need for management of the bank, which can
handle all this with comfort and ease.
Smooth and efficient management affects the satisfaction of the customers and
staff members, indirectly. And of course, it encourages management committee in
taking some needed decision for future enhancement of the bank.
Now a day’s, managing a bank is tedious job up to certain limit. So software that
reduces the work is essential. Also, today’s world is a genuine computer world and
is getting faster and faster day-by-day. Thus, considering above necessities, the
software for bank management has become necessary which would be useful in
managing the bank more efficiently.
All transactions are carried out online by transferring from accounts in the same
Bank or international bank. The software is meant to overcome the drawbacks of
the manual system.
7
• AIM of this project
The main aim of designing and developing this Internet banking System Java
primarily based Engineering project is to provide secure and efficient net banking
facilities to the banking customers over the internet. Apache Server Pages,
MYSQL database used to develop this bank application where all banking
customers can login through the secured web page by their account login id and
password. Users will have all options and features in that application like get
money from western union, money transfer to others, and send cash or money to
inter banking as well as other banking customers by simply adding them as
payees.
• Main Purpose
 The Traditional way of maintaining details of a user in a bank was to
enter the details and record them. Every time the user needs to perform some
transactions he has to go to bank and perform the necessary actions, which may
not be so feasible all the time. It may be a hard-hitting task for the users and the
bankers too. The project gives real life understanding of Online Banking System
and activities performed by various roles in the supply chain. Here, we provide
automation for banking system through Internet. Online Banking System project
captures activities performed by different roles in real life banking which provides
enhanced techniques for maintaining the required information up-to-date, which
results in efficiency. The project gives real life understanding of Online Banking
System and activities performed by various roles in the supply chain
• Main Goal
1. Motto- Our motto is to develop a software program for managing the entire
bank process related to Administration accounts customer accounts and to keep
each every track about their property and their various transaction processes
efficiently.
Hereby, our main objective is the customer’s satisfaction considering today’s
faster in the world.
2. Customer Satisfaction: Client can do his operations comfortably without any
risk or losing of his privacy. Our software will perform and fulfill all the tasks that
any customer would desire.
3. Saving Customer Time: Client doesn't need to go to the bank to do small
operation.
4. Protecting the Customer: It helps the customer to be satisfied and comfortable
in his choices, this protection contains customer’s account, money and his rivacy.
5. Transferring Money: Help client transferring money to/or another bank or
country